Introduce class Kitchen and Kitchen.loadIngredients
This class has a static function loadIngredients

This function takes in a list of ingridients and a reference date
object tha represents the current time.

It will then loop through the ingredients and separate the ingredients
into two based on useBy date. (expired vs unexpired)

Unexpired ingridients will further have a value (bestBeforeIndex)
that indicates how far from or beyond we are from the
best before date of the refenced ingredient.