Introduce Kitchen.prepareRecipes and Kitchen constructor
This static method takes in the result of Kitchen.loadIngredients
together with a second param that is a list of recipes.

It will then loop through all recipes and compared the ingredient
list with our list on available(unexpired) ingredients.

A recipe is marked as available if all ingredients are available
otherwise it is marked as unavailable.

This function will then return two arrays containing available and
unavailable recipes.

We provide a constructor that stitches up Kitchen.loadIngredients
and Kitchen.prepareRecipes.

We provide a class method `recommendARecipe` that returns the available
recipes.